requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
textblob==0.17.1
aiofiles==23.2.0
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0 